Remove Plots weak dep and fix DataPipeline and PyCall
